requirements.txt
#pip freeze > requirements.txt
#获取目录下的所有依赖项

#如果只需要收集用到的依赖项
#需要先安装pip install pipreqs
#pipreqs . --encoding=utf8 --force


# Using venv (Python 3.3+)
#创建虚拟环境python -m venv myenv

# Activate the virtual environment
#激活虚拟环境
# On Windows:
#myenv\Scripts\activate

gTTS==2.3.2
openai==0.27.8
playsound==1.3.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
streamlit==1.25.0
speech-recognition-python==3.9.9
PyAudio==0.2.13